修正途中

+++ src/main/java/com/kyouseipro/neo/qualification/entity/QualificationsEntity.java
@@ -9,7 +9,7 @@ import lombok.Data;
 
 @Data
 public class QualificationsEntity implements CsvExportable {
-    private int qualificationsId;
+    private Long qualificationsId;
     private Integer employeeId;
     private Integer companyId;
     private String ownerName;
